Reimplement updating of segger settings
The previous implementation only handled the case that there already
was a segger configuration with just potentially unfitting nrf settings
in it. The new implementation handles two more cases:

1. If there is no settings file at all yet (which will be the case if
   the users had never installed or started segger before).
2. If there is a settings file but without the sought nrf settings in it
   (which will be the case if the users never configured them manually).

This also adds tests for updating the settings. This test code is just
disabled until we update jest in the shared project.

Also this new implementation uses the browser build in DOMParser and
XMLSerializer, so we do not need xml2js anymore.