Reflect a partially deleted environment in the UI
In some cases (e.g. if a bash is still open) the `move(srcDir,
toBeDeletedDir)` succeeds but the following `move(srcDir,
toBeDeletedDir)` not. Then the UI still showed the SDK as being
installed, while the folder was alreade renamed (and mostly deleted). In
this state the user could not interact with the environment anymore and
instead had to reload the toolchain manager app.

Now in that case still an error is shown, but the UI reflects the new
state: If the environment is available online it will again be shown as
installable and if existed only locally, then it is removed.