Make different states more clear
An environment can be in a certain state (availableForDownload,
installing, cloning, installed or removing). This was rather mixed up
before and is hopefully more clearly expressed now.

While doing so, a lot of selectors were also added to the
environmentReducer, so that the logic how to retrieve what information
from an environment is not spread out and duplicated all over the place.