Fix just updating the NCS
The previous implementation sometimes corrupted the installation. The
new implementation does respect whether a new NCS is installed or
whether it is just updated. In the latter case, a possibly existing
directory `.west` must not be removed and the script `ncsmgr init-ncs`
should not call `west init`. For this, the new implementation requires
an updated version of the toolchain, that honors passing the parameter
`--just-update` to the script `ncsmgr init-ncs`. This update is
implemented in https://github.com/NordicPlayground/toolchain/pull/26.

This new implementation follows the logic as described in
https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/nrf/gs_installing.html#updating-the-repositories