Fix path on macOS
Strangely, /usr/local/bin is not in the PATH by default when we open
a terminal from an application. But it being in the path is expected
for many cases, e.g. when nrfjprog is placed there. homebrew is usually
also found in /usr/local/bin

(cherry picked from commit 3f77fde70dcd832be73ece544eb2473c8bf738eb)