Fix wrong display “null 0%”
While installing a “null 0%” shortly flashed up after the toolchain was
unpacked when cloning the SDK started.

This was caused by the fact that we set `stage` to `null` (according to
the type only `undefined`, so I changed it to now also use undefined) is
allowed. And we only checked for `undefined` (as a second measurement I
also changed that to now check for both, `null` and `undefined.)